[ROS][asyncros] Split code
