+++ ros2/amiga_control/amiga_control/aioros2/decorators/param.py
@@ -1,0 +1,8 @@
+from ._decorators import decorate_handler, ros_type_e
+
+def param(dataclass_param):
+    def _param(fn):
+        decorate_handler(fn, ros_type_e.PARAM)
+        return fn
+
+    return _param
