+++ ros2/amiga_control/amiga_control/aioros2/feedback.py
@@ -1,0 +1,6 @@
+from collections import namedtuple
+
+Feedback = namedtuple("Feedback", ["args", "kwargs"])
+
+def feedback(*args, **kwargs):
+    return Feedback(args, kwargs)
