+++ ros2/amiga_control/amiga_control/aioros2/result.py
@@ -1,0 +1,6 @@
+from collections import namedtuple
+
+Result = namedtuple("Respond", ["args", "kwargs"])
+
+def result(*args, **kwargs):
+    return Result(args, kwargs)
